feat: centralize server entry point
