+++ package.json
@@ -17,7 +17,7 @@
     "migrate": "tsx scripts/migrate.ts",
     "api-dev": "tsx watch src/index.ts",
     "build:api": "tsc -p tsconfig.api.json",
-    "api": "npm run build:api && node dist/src/index.js",
+    "api": "npm run build:api && node dist/index.js",
     "lighthouse": "lhci autorun",
     "audit": "npm audit --audit-level=moderate",
     "analyze": "ANALYZE=true next build"
